// src/app/interfaces/ProductAllergens.ts
export interface ProductAllergens {
    id: number;
    productName: string;
    gluten: boolean;
    crustaceos: boolean;
    huevos: boolean;
    pescado: boolean;
    cacahuetes: boolean;
    soja: boolean;
    lacteos: boolean;
    frutosSecos: boolean;
    apio: boolean;
    mostaza: boolean;
    sesamo: boolean;
    sulfitos: boolean;
    altramuz: boolean;
    moluscos: boolean;

}/// END OF INTERFACE GamePlayed //